+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- New `page_header` component.
 - Added `gettext` attribute to `Doggo.input/1`, giving you the choice to
   set the Gettext module locally.